+++ src/cli/mapshaper-option-validation.js
@@ -74,9 +74,11 @@ function validateProjOpts(cmd) {
   if (_.length > 0) {
     error("Received one or more unexpected parameters: " + _.join(', '));
   }
-  if (!cmd.options.projection && !cmd.options.source) {
-    error("Missing projection data");
+
+  if (!(cmd.options.projection  || cmd.options.match || cmd.options.from)) {
+    stop("Missing projection data");
   }
+
 }
 
 function validateJoinOpts(cmd) {